Move rename, astype
